+ changed addNode, so that the parent node will get the node as children. + Added Categories to CupidMatcher.java + Added raw Construct for match(Tablepair tablepair) method.

+++ src/main/java/de/uni_marburg/schematch/matching/sota/cupid/StructuralSimilarity.java
@@ -7,7 +7,7 @@ public class StructuralSimilarity {
             SchemaElementNode s,
             SchemaElementNode t,
             Map<String, Map<StringPair, Float>> sims,
-            double th_accept
+            float th_accept
     ) {
         List<SchemaElementNode> sLeaves = s.leaves();
         List<SchemaElementNode> tLeaves = t.leaves();
